+++ crates/tanren-app-services/Cargo.toml
@@ -18,6 +18,5 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 tanren-contract = { path = "../tanren-contract" }
 tanren-identity-policy = { path = "../tanren-identity-policy" }
-tanren-policy = { path = "../tanren-policy" }
 tanren-store = { path = "../tanren-store" }
 thiserror = { workspace = true }
