+++ crates/tanren-policy/src/lib.rs
@@ -5,7 +5,6 @@
 //! [`Decision`] enum below.
 
 use serde::{Deserialize, Serialize};
-use tanren_identity_policy::OrganizationPermission;
 use thiserror::Error;
 
 /// The outcome of evaluating a policy against an actor and a resource.
@@ -39,70 +38,4 @@ pub enum PolicyError {
     /// Required policy inputs were missing or malformed.
     #[error("policy evaluation failed: missing input '{0}'")]
     MissingInput(String),
-}
-
-/// Stable capability metadata for organization-scoped permission gates.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct OrganizationCapability {
-    /// Stable capability id surfaced to interface layers.
-    pub key: &'static str,
-    /// Human-readable capability summary.
-    pub summary: &'static str,
-}
-
-/// Permission gate definition owned by policy.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct OrganizationPermissionGate {
-    /// Capability metadata for this gate.
-    pub capability: OrganizationCapability,
-    /// Backing organization permission required by the capability.
-    pub permission: OrganizationPermission,
-}
-
-impl OrganizationPermissionGate {
-    /// Build a policy-owned gate from an organization permission.
-    #[must_use]
-    pub const fn from_permission(permission: OrganizationPermission) -> Self {
-        Self {
-            capability: organization_capability(permission),
-            permission,
-        }
-    }
-}
-
-/// Map an organization permission to stable capability metadata.
-#[must_use]
-pub const fn organization_capability(permission: OrganizationPermission) -> OrganizationCapability {
-    match permission {
-        OrganizationPermission::Invite => OrganizationCapability {
-            key: "organization.invite",
-            summary: "Invite people to the organization",
-        },
-        OrganizationPermission::ManageAccess => OrganizationCapability {
-            key: "organization.manage_access",
-            summary: "Manage organization access for members",
-        },
-        OrganizationPermission::Configure => OrganizationCapability {
-            key: "organization.configure",
-            summary: "Configure organization-level defaults",
-        },
-        OrganizationPermission::SetPolicy => OrganizationCapability {
-            key: "organization.set_policy",
-            summary: "Manage organization policy settings",
-        },
-        OrganizationPermission::Delete => OrganizationCapability {
-            key: "organization.delete",
-            summary: "Delete the organization",
-        },
-    }
-}
-
-/// Evaluate a permission gate from a concrete store permission result.
-#[must_use]
-pub const fn evaluate_organization_permission_gate(allowed: bool) -> Decision {
-    if allowed {
-        Decision::Allow
-    } else {
-        Decision::Deny(DenialReason::MissingPermission)
-    }
 }
